Setting up parallel branch

+++ DHSVM3.1.2/sourcecode/SnowMelt.c
@@ -1,6 +1,6 @@
 /*
  * SUMMARY:      SnowMelt.c - Calculate snow accumulation and melt
- * USAGE:        
+ * USAGE:
  *
  * AUTHOR:       Mark Wigmosta and Pascal Storck
  * ORG:          University of Washington, Department of Civil Engineering
@@ -11,22 +11,22 @@
  * DESCRIP-END.
  * FUNCTIONS:    SnowMelt()
  * COMMENTS:
- * $Id: SnowMelt.c,v 1.4 2003/07/01 21:26:25 olivier Exp $     
+ * $Id: SnowMelt.c,v 1.4 2003/07/01 21:26:25 olivier Exp $
  */
 
+#include "brent.h"
+#include "constants.h"
+#include "functions.h"
+#include "massenergy.h"
+#include "settings.h"
+#include "snow.h"
 #include <assert.h>
 #include <math.h>
 #include <stdarg.h>
 #include <stdio.h>
 #include <stdlib.h>
-#include "brent.h"
-#include "constants.h"
-#include "settings.h"
-#include "massenergy.h"
-#include "functions.h"
-#include "snow.h"
 
-static float CalcSnowPackEnergyBalance(float Tsurf, ...);
+static float CalcSnowPackEnergyBalance(double Tsurf, ...);
 
 /*****************************************************************************
   Function name: SnowMelt()
@@ -38,13 +38,13 @@ static float CalcSnowPackEnergyBalance(float Tsurf, ...);
     int y                  - Row counter
     int x                  - Column counter
     int Dt                 - Model timestep (seconds)
-    float Z                - Reference height (m) 
+    float Z                - Reference height (m)
     float Displacement     - Displacement height (m)
     float Z0               - Surface roughness (m)
     float BaseRa           - Aerodynamic resistance (uncorrected for
                              stability) (s/m)
     float AirDens          - Density of air (kg/m3)
-    float EactAir          - Actual vapor pressure of air (Pa) 
+    float EactAir          - Actual vapor pressure of air (Pa)
     float Lv               - Latent heat of vaporization (J/kg3)
     float ShortRad         - Net exchange of shortwave radiation (W/m2)
     float LongRadIn        - Incoming long wave radiation (W/m2)
@@ -54,8 +54,8 @@ static float CalcSnowPackEnergyBalance(float Tsurf, ...);
     float Tair             - Air temperature (C)
     float Vpd              - Vapor pressure deficit (Pa)
     float Wind             - Wind speed (m/s)
-    float *PackWater       - Liquid water content of snow pack 
-    float *SurfWater	   - Liquid water content of surface layer 
+    float *PackWater       - Liquid water content of snow pack
+    float *SurfWater	   - Liquid water content of surface layer
     float *Swq             - Snow water equivalent at current pixel (m)
     float *VaporMassFlux;  - Mass flux of water vapor to or from the
                              intercepted snow (m)
@@ -68,8 +68,8 @@ static float CalcSnowPackEnergyBalance(float Tsurf, ...);
     float Outflow          - Amount of snowpack outflow (m)
 
   Modifies     :
-    float *PackWater       - Liquid water content of snow pack 
-    float *SurfWater	   - Liquid water content of surface layer 
+    float *PackWater       - Liquid water content of snow pack
+    float *SurfWater	   - Liquid water content of surface layer
     float *Swq             - Snow water equivalent at current pixel (m)
     float *VaporMassFlux;  - Mass flux of water vapor to or from the
                              intercepted snow (m)
@@ -81,33 +81,32 @@ static float CalcSnowPackEnergyBalance(float Tsurf, ...);
   Comments     :
 *****************************************************************************/
 float SnowMelt(int y, int x, int Dt, float Z, float Displacement, float Z0,
-	       float BaseRa, float AirDens, float EactAir, float Lv,
-	       float ShortRad, float LongRadIn, float Press, float RainFall,
-	       float SnowFall, float Tair, float Vpd, float Wind,
-	       float *PackWater, float *SurfWater, float *Swq,
-	       float *VaporMassFlux, float *TPack, float *TSurf,
-	       float *MeltEnergy)
-{
-  float DeltaPackCC;		/* Change in cold content of the pack */
-  float DeltaPackSwq;		/* Change in snow water equivalent of the
-				   pack (m) */
-  float Ice;			/* Ice content of snow pack (m) */
-  float InitialSwq;		/* Initial snow water equivalent (m) */
-  float MassBalanceError;	/* Mass balance error (m) */
-  float MaxLiquidWater;		/* Maximum liquid water content of pack (m) */
-  float OldTSurf;		/* Old snow surface temperature (C) */
-  float Outflow;		/* Amount water flowing out of the snow pack
-				   during the time interval (m) */
-  float PackCC;			/* Cold content of snow pack (J) */
-  float PackSwq;		/* Snow pack snow water equivalent (m) */
-  float Qnet;			/* Net energy exchange at the surface (W/m2) */
-  float RefreezeEnergy;		/* refreeze energy (W/m2) */
-  float RefrozenWater;		/* Amount of refrozen water (m) */
-  float SnowFallCC;		/* Cold content of new snowfall (J) */
-  float SnowMelt;		/* Amount of snow melt during time interval
-				   (m water equivalent) */
-  float SurfaceCC;		/* Cold content of snow pack (J) */
-  float SurfaceSwq;		/* Surface layer snow water equivalent (m) */
+               float BaseRa, float AirDens, float EactAir, float Lv,
+               float ShortRad, float LongRadIn, float Press, float RainFall,
+               float SnowFall, float Tair, float Vpd, float Wind,
+               float *PackWater, float *SurfWater, float *Swq,
+               float *VaporMassFlux, float *TPack, float *TSurf,
+               float *MeltEnergy) {
+  float DeltaPackCC;      /* Change in cold content of the pack */
+  float DeltaPackSwq;     /* Change in snow water equivalent of the
+                             pack (m) */
+  float Ice;              /* Ice content of snow pack (m) */
+  float InitialSwq;       /* Initial snow water equivalent (m) */
+  float MassBalanceError; /* Mass balance error (m) */
+  float MaxLiquidWater;   /* Maximum liquid water content of pack (m) */
+  float OldTSurf;         /* Old snow surface temperature (C) */
+  float Outflow;          /* Amount water flowing out of the snow pack
+                             during the time interval (m) */
+  float PackCC;           /* Cold content of snow pack (J) */
+  float PackSwq;          /* Snow pack snow water equivalent (m) */
+  float Qnet;             /* Net energy exchange at the surface (W/m2) */
+  float RefreezeEnergy;   /* refreeze energy (W/m2) */
+  float RefrozenWater;    /* Amount of refrozen water (m) */
+  float SnowFallCC;       /* Cold content of new snowfall (J) */
+  float SnowMelt;         /* Amount of snow melt during time interval
+                             (m water equivalent) */
+  float SurfaceCC;        /* Cold content of snow pack (J) */
+  float SurfaceSwq;       /* Surface layer snow water equivalent (m) */
 
   InitialSwq = *Swq;
   OldTSurf = *TSurf;
@@ -135,16 +134,15 @@ float SnowMelt(int y, int x, int Dt, float Z, float Displacement, float Z0,
   if (SnowFall > (MAX_SURFACE_SWE - SurfaceSwq)) {
     DeltaPackSwq = SurfaceSwq + SnowFall - MAX_SURFACE_SWE;
     if (DeltaPackSwq > SurfaceSwq)
-      DeltaPackCC = SurfaceCC + (SnowFall - MAX_SURFACE_SWE) / SnowFall *
-	SnowFallCC;
+      DeltaPackCC =
+          SurfaceCC + (SnowFall - MAX_SURFACE_SWE) / SnowFall * SnowFallCC;
     else
       DeltaPackCC = DeltaPackSwq / SurfaceSwq * SurfaceCC;
     SurfaceSwq = MAX_SURFACE_SWE;
     SurfaceCC += SnowFallCC - DeltaPackCC;
     PackSwq += DeltaPackSwq;
     PackCC += DeltaPackCC;
-  }
-  else {
+  } else {
     SurfaceSwq += SnowFall;
     SurfaceCC += SnowFallCC;
   }
@@ -163,11 +161,10 @@ float SnowMelt(int y, int x, int Dt, float Z, float Displacement, float Z0,
 
   /* Calculate the surface energy balance for snow_temp = 0.0 */
 
-  Qnet = CalcSnowPackEnergyBalance((float) 0.0, Dt, BaseRa, Z, Displacement,
-				   Z0, Wind, ShortRad, LongRadIn, AirDens,
-				   Lv, Tair, Press, Vpd, EactAir, RainFall,
-				   SurfaceSwq, *SurfWater, OldTSurf,
-				   &RefreezeEnergy, VaporMassFlux);
+  Qnet = CalcSnowPackEnergyBalance(
+      (float)0.0, Dt, BaseRa, Z, Displacement, Z0, Wind, ShortRad, LongRadIn,
+      AirDens, Lv, Tair, Press, Vpd, EactAir, RainFall, SurfaceSwq, *SurfWater,
+      OldTSurf, &RefreezeEnergy, VaporMassFlux);
 
   /* If Qnet == 0.0, then set the surface temperature to 0.0 */
   if (fequal(Qnet, 0.0)) {
@@ -175,8 +172,8 @@ float SnowMelt(int y, int x, int Dt, float Z, float Displacement, float Z0,
     if (RefreezeEnergy >= 0.0) {
       RefrozenWater = RefreezeEnergy / (LF * WATER_DENSITY) * Dt;
       if (RefrozenWater > *SurfWater) {
-	RefrozenWater = *SurfWater;
-	RefreezeEnergy = (RefrozenWater * LF * WATER_DENSITY) / Dt;
+        RefrozenWater = *SurfWater;
+        RefreezeEnergy = (RefrozenWater * LF * WATER_DENSITY) / Dt;
       }
       *MeltEnergy += RefreezeEnergy;
       SurfaceSwq += RefrozenWater;
@@ -185,8 +182,7 @@ float SnowMelt(int y, int x, int Dt, float Z, float Displacement, float Z0,
       assert(*SurfWater >= 0.0);
       SnowMelt = 0.0;
 
-    }
-    else {
+    } else {
 
       /* Calculate snow melt */
       SnowMelt = fabs(RefreezeEnergy) / (LF * WATER_DENSITY) * Dt;
@@ -199,24 +195,22 @@ float SnowMelt(int y, int x, int Dt, float Z, float Displacement, float Z0,
     if (*SurfWater < -(*VaporMassFlux)) {
       *VaporMassFlux = -(*SurfWater);
       *SurfWater = 0.0;
-    }
-    else
+    } else
       *SurfWater += *VaporMassFlux;
 
     /* If SnowMelt < Ice, there was incomplete melting of the pack */
 
     if (SnowMelt < Ice) {
       if (SnowMelt <= PackSwq) {
-	*SurfWater += SnowMelt;
-	PackSwq -= SnowMelt;
-	Ice -= SnowMelt;
-      }
-      else {
-	*SurfWater += SnowMelt + *PackWater;
-	*PackWater = 0.0;
-	PackSwq = 0.0;
-	Ice -= SnowMelt;
-	SurfaceSwq = Ice;
+        *SurfWater += SnowMelt;
+        PackSwq -= SnowMelt;
+        Ice -= SnowMelt;
+      } else {
+        *SurfWater += SnowMelt + *PackWater;
+        *PackWater = 0.0;
+        PackSwq = 0.0;
+        Ice -= SnowMelt;
+        SurfaceSwq = Ice;
       }
     }
 
@@ -236,11 +230,11 @@ float SnowMelt(int y, int x, int Dt, float Z, float Displacement, float Z0,
   else {
     /* Calculate surface layer temperature using "Brent method" */
 
-    *TSurf = RootBrent(y, x, (float) (*TSurf - DELTAT), (float) 0.0,
-		       SnowPackEnergyBalance, Dt, BaseRa, Z, Displacement,
-		       Z0, Wind, ShortRad, LongRadIn, AirDens, Lv, Tair,
-		       Press, Vpd, EactAir, RainFall, SurfaceSwq, *SurfWater,
-		       OldTSurf, &RefreezeEnergy, VaporMassFlux);
+    *TSurf = RootBrent(y, x, (float)(*TSurf - DELTAT), (float)0.0,
+                       SnowPackEnergyBalance, Dt, BaseRa, Z, Displacement, Z0,
+                       Wind, ShortRad, LongRadIn, AirDens, Lv, Tair, Press, Vpd,
+                       EactAir, RainFall, SurfaceSwq, *SurfWater, OldTSurf,
+                       &RefreezeEnergy, VaporMassFlux);
 
     /* since we iterated, the surface layer is below freezing and no snowmelt
      */
@@ -263,8 +257,7 @@ float SnowMelt(int y, int x, int Dt, float Z, float Displacement, float Z0,
       *VaporMassFlux = -SurfaceSwq;
       SurfaceSwq = 0.0;
       Ice = PackSwq;
-    }
-    else {
+    } else {
       SurfaceSwq += *VaporMassFlux;
       Ice += *VaporMassFlux;
     }
@@ -277,38 +270,36 @@ float SnowMelt(int y, int x, int Dt, float Z, float Displacement, float Z0,
   if (*SurfWater > MaxLiquidWater) {
     Outflow = *SurfWater - MaxLiquidWater;
     *SurfWater = MaxLiquidWater;
-  }
-  else
+  } else
     Outflow = 0.0;
 
-  /* Refreeze liquid water in the pack.                                   
-     variable 'RefreezeEnergy' is the heat released to the snow pack            
-     if all liquid water were refrozen.                                   
-     if RefreezeEnergy < PackCC then all water IS refrozen           
-     PackCC always <=0.0 
+  /* Refreeze liquid water in the pack.
+     variable 'RefreezeEnergy' is the heat released to the snow pack
+     if all liquid water were refrozen.
+     if RefreezeEnergy < PackCC then all water IS refrozen
+     PackCC always <=0.0
 
-     WORK IN PROGRESS: This energy is NOT added to MeltEnergy, since this does 
-     not involve energy transported to the pixel.  Instead heat from the snow 
+     WORK IN PROGRESS: This energy is NOT added to MeltEnergy, since this does
+     not involve energy transported to the pixel.  Instead heat from the snow
      pack is used to refreeze water */
 
-  *PackWater += Outflow;	/* add surface layer outflow to pack liquid water */
+  *PackWater += Outflow; /* add surface layer outflow to pack liquid water */
   RefreezeEnergy = *PackWater * LF * WATER_DENSITY;
 
   /* calculate energy released to freeze */
 
-  if (PackCC < -RefreezeEnergy) {	/* cold content not fully depleted */
-    PackSwq += *PackWater;	/* refreeze all water and update */
+  if (PackCC < -RefreezeEnergy) { /* cold content not fully depleted */
+    PackSwq += *PackWater;        /* refreeze all water and update */
     Ice += *PackWater;
     *PackWater = 0.0;
     if (PackSwq > 0.0)
       *TPack = (PackCC + RefreezeEnergy) / (CH_ICE * PackSwq);
     else
       *TPack = 0.0;
-  }
-  else {
+  } else {
     /* cold content has been either exactly satisfied or exceeded. If
        PackCC = refreeze then pack is ripe and all pack water is
-       refrozen, else if energy released in refreezing exceeds PackCC 
+       refrozen, else if energy released in refreezing exceeds PackCC
        then exactly the right amount of water is refrozen to satify PackCC.
        The refrozen water is added to PackSwq and Ice */
 
@@ -325,8 +316,7 @@ float SnowMelt(int y, int x, int Dt, float Z, float Displacement, float Z0,
   if (*PackWater > MaxLiquidWater) {
     Outflow = *PackWater - MaxLiquidWater;
     *PackWater = MaxLiquidWater;
-  }
-  else
+  } else
     Outflow = 0.0;
 
   /* Update snow properties */
@@ -341,8 +331,7 @@ float SnowMelt(int y, int x, int Dt, float Z, float Displacement, float Z0,
       SurfaceCC -= SurfaceCC * (SurfaceSwq - MAX_SURFACE_SWE) / SurfaceSwq;
       PackSwq += SurfaceSwq - MAX_SURFACE_SWE;
       SurfaceSwq -= SurfaceSwq - MAX_SURFACE_SWE;
-    }
-    else if (SurfaceSwq < MAX_SURFACE_SWE) {
+    } else if (SurfaceSwq < MAX_SURFACE_SWE) {
       PackCC -= PackCC * (MAX_SURFACE_SWE - SurfaceSwq) / PackSwq;
       SurfaceCC += PackCC * (MAX_SURFACE_SWE - SurfaceSwq) / PackSwq;
       PackSwq -= MAX_SURFACE_SWE - SurfaceSwq;
@@ -350,8 +339,7 @@ float SnowMelt(int y, int x, int Dt, float Z, float Displacement, float Z0,
     }
     *TPack = PackCC / (CH_ICE * PackSwq);
     *TSurf = SurfaceCC / (CH_ICE * SurfaceSwq);
-  }
-  else {
+  } else {
     PackSwq = 0.0;
     PackCC = 0.0;
     *TPack = 0.0;
@@ -366,8 +354,8 @@ float SnowMelt(int y, int x, int Dt, float Z, float Displacement, float Z0,
 
   /* Mass balance test */
 
-  MassBalanceError = (InitialSwq - *Swq) + (RainFall + SnowFall) - Outflow +
-    *VaporMassFlux;
+  MassBalanceError =
+      (InitialSwq - *Swq) + (RainFall + SnowFall) - Outflow + *VaporMassFlux;
 
   return (Outflow);
 }
@@ -378,7 +366,7 @@ float SnowMelt(int y, int x, int Dt, float Z, float Displacement, float Z0,
   Purpose      : Dummy function to make a direct call to
                  SnowEnergyBalance() possible.
 
-  Required     : 
+  Required     :
     float TSurf - SnowPack surface temperature (C)
     other arguments required by SnowPackEnergyBalance()
 
@@ -389,12 +377,11 @@ float SnowMelt(int y, int x, int Dt, float Z, float Displacement, float Z0,
 
   Comments     : function is local to this module
 *****************************************************************************/
-static float CalcSnowPackEnergyBalance(float Tsurf, ...)
-{
-  va_list ap;			/* Used in traversing variable argument list
-				 */
-  float Qnet;			/* Net energy exchange at the SnowPack snow
-				   surface (W/m^2) */
+static float CalcSnowPackEnergyBalance(double Tsurf, ...) {
+  va_list ap; /* Used in traversing variable argument list
+               */
+  float Qnet; /* Net energy exchange at the SnowPack snow
+                 surface (W/m^2) */
 
   va_start(ap, Tsurf);
   Qnet = SnowPackEnergyBalance(Tsurf, ap);
